adding readme, fixing value2 parseFloat

+++ main.js
@@ -1,5 +1,5 @@
 function sueldoPorHoras() {
 	var num1 = parseInt(document.getElementById('n1').value);
-	var num2 = parseInt(document.getElementById('n2').value);
+	var num2 = parseFloat(document.getElementById('n2').value);
     document.getElementById('resultado').value = num1 * num2;
 }
